askama

Module filters

Module for built-in filter functions

Contains all the built-in filter functions for use in templates. You can define your own filters, as well.

Note

All result types of any filter function in this module is subject to change at any point, and is not indicated by as semver breaking version bump. The traits AutoEscape and WriteWritable are used by askama_macros’s generated code to work with all compatible types.

Structs

AutoEscaper

Used internally by askama to select the appropriate escaper

Html

Escape characters in a safe way for HTML texts and attributes

HtmlSafeOutput

Like Safe, but only for HTML output

Safe

Mark the output of a filter as “safe”

Text

Don’t escape the input but return in verbatim

Unsafe

There is not need to mark the output of a custom filter as “unsafe”; this is simply the default

Writable

Used internally by askama to select the appropriate write!() mechanism

Enums

Either

Render either L or R

MaybeSafe

Mark the output of a filter as “maybe safe”

Traits

AsIndent

A prefix usable for indenting |indent.

AutoEscape

Used internally by askama to select the appropriate escaper

DefaultFilterable

A type (or a reference to it) that can be used in |assigned_or.

Escaper

Escapers are used to make generated text safe for printing in some context.

HtmlSafe

Types that implement this marker trait don’t need to be HTML escaped

PluralizeCount

An integer that can have the value +1 and maybe -1.

WriteWritable

Used internally by askama to select the appropriate write!() mechanism

Functions

assigned_or

Render value if it is not its “default” value, see DefaultFilterable, otherwise fallback.

capitalize

Capitalize a value. The first character will be uppercase, all others lowercase.

center

Centers the value in a field of a given width

e

Alias for escape()

escape

Escapes strings according to the escape mode.

filesizeformat

Returns adequate string representation (in KB, ..) of number of bytes

fmt

Formats arguments according to the specified format

format

Formats arguments according to the specified format

indent

Indent lines with spaces or a prefix.

join

Joins iterable into a string separated by provided argument

linebreaks

Replaces line breaks in plain text with appropriate HTML.

linebreaksbr

Converts all newlines in a piece of plain text to HTML line breaks.

lower

Converts to lowercase

lowercase

Converts to lowercase, alias for the |lower filter

paragraphbreaks

Replaces only paragraph breaks in plain text with appropriate HTML

pluralize

For a value of ±1 by default an empty string "" is returned, otherwise "s".

reject

Returns an iterator without filtered out values.

reject_with

Returns an iterator without filtered out values.

safe

Marks a string (or other Display type) as safe

title

Return a title cased version of the value. Words will start with uppercase letters, all remaining characters are lowercase.

titlecase

Return a title cased version of the value. Alias for the |title filter.

trim

Strip leading and trailing whitespace

truncate

Limit string length, appends ‘…’ if truncated

unique

Returns an iterator with all duplicates removed.

upper

Converts to uppercase

uppercase

Converts to uppercase, alias for the |upper filter

urlencode

Percent-encodes the argument for safe use in URI; does not encode /.

urlencode_strict

Percent-encodes the argument for safe use in URI; encodes /.

wordcount

Count the words in that string.