Function askama::filters::paragraphbreaks

pub fn paragraphbreaks<T>(s: T) -> Result<String, Error>where
    T: Display,

Replaces only paragraph breaks in plain text with appropriate HTML

A new line followed by a blank line becomes a paragraph break <p>. Paragraph tags only wrap content; empty paragraphs are removed. No <br/> tags are added.