askama::filters

Function safe

pub fn safe<T, E>(text: T, escaper: E) -> Result<Safe<T>, Infallible>

Marks a string (or other Display type) as safe

Use this if you want to allow markup in an expression, or if you know that the expression’s contents don’t need to be escaped.

Askama will automatically insert the first (Escaper) argument, so this filter only takes a single argument of any type that implements Display.

/// ```jinja
/// <div>{{ example|safe }}</div>
/// ```
#[derive(Template)]
#[template(ext = "html", in_doc = true)]
struct Example<'a> {
    example: &'a str,
}

assert_eq!(
    Example { example: "<p>I'm Safe</p>" }.to_string(),
    "<div><p>I'm Safe</p></div>"
);