askama::filters

Trait DefaultFilterable

pub trait DefaultFilterable {
    type Filtered<'a>
       where Self: 'a;
    type Error: Into<Error>;

    // Required method
    fn as_filtered(&self) -> Result<Option<Self::Filtered<'_>>, Self::Error>;
}

A type (or a reference to it) that can be used in |assigned_or.

The type is either a monad such as Option or Result, or a type that has a well defined, trivial default value, e.g. an empty str or 0 for integer types.

Required Associated Types

type Filtered<'a> where Self: 'a

The contained value

type Error: Into<Error>

An error that prevented as_filtered() to succeed, e.g. a poisoned state or an unacquirable lock.

Required Methods

fn as_filtered(&self) -> Result<Option<Self::Filtered<'_>>, Self::Error>

Return the contained value, if a value was contained, and it’s not the default value.

Returns Ok(None) if the value could not be unwrapped.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.