askama::filters

Trait PluralizeCount

pub trait PluralizeCount {
    type Error: Into<Error>;

    // Required method
    fn is_singular(&self) -> Result<bool, Self::Error>;
}

An integer that can have the value +1 and maybe -1.

Required Associated Types

type Error: Into<Error>

A possible error that can occur while checking the value.

Required Methods

fn is_singular(&self) -> Result<bool, Self::Error>

Returns true if and only if the value is ±1.