askama::filters

Trait WriteWritable

pub trait WriteWritable {
    // Required method
    fn askama_write(
        &self,
        dest: &mut dyn Write,
        values: &dyn Values,
    ) -> Result<()>;
}

Used internally by askama to select the appropriate write!() mechanism

Required Methods

fn askama_write(&self, dest: &mut dyn Write, values: &dyn Values) -> Result<()>

Used internally by askama to select the appropriate write!() mechanism