askama

Trait FastWritable

pub trait FastWritable {
    // Required method
    fn write_into(
        &self,
        dest: &mut dyn Write,
        values: &dyn Values,
    ) -> Result<()>;
}

Types implementing this trait can be written without needing to employ an fmt::Formatter.

Required Methods

fn write_into(&self, dest: &mut dyn Write, values: &dyn Values) -> Result<()>

Used internally by askama to speed up writing some types.