Module askama::filters

Module for built-in filter functions

Contains all the built-in filter functions for use in templates. You can define your own filters; for more information, see the top-level crate documentation.

Constants

BUILT_IN_FILTERS

Functions

abs
Absolute value

capitalize
Capitalize a value. The first character will be uppercase, all others lowercase.

center
Centers the value in a field of a given width

e
Alias for the escape() filter

escape
Escapes &, < and > in strings

format
Formats arguments according to the specified format

join
Joins iterable into a string separated by provided argument

linebreaks
Replaces line breaks in plain text with appropriate HTML

linebreaksbr
Converts all newlines in a piece of plain text to HTML line breaks

lower
Converts to lowercase

lowercase
Alias for the lower() filter

safe
Marks a string (or other Display type) as safe

trim
Strip leading and trailing whitespace

truncate
Limit string length, appends '...' if truncated

upper
Converts to uppercase

uppercase
Alias for the upper() filter

wordcount
Count the words in that string