askama::filters

Trait AutoEscape

pub trait AutoEscape {
    type Escaped: Display;
    type Error: Into<Error>;

    // Required method
    fn askama_auto_escape(&self) -> Result<Self::Escaped, Self::Error>;
}

Used internally by askama to select the appropriate escaper

Required Associated Types

type Escaped: Display

The wrapped or converted result type

type Error: Into<Error>

Early error testing for the input value, usually Infallible

Required Methods

fn askama_auto_escape(&self) -> Result<Self::Escaped, Self::Error>

Used internally by askama to select the appropriate escaper