askama::filters

Trait HtmlSafe

pub trait HtmlSafe: Display { }

Types that implement this marker trait don’t need to be HTML escaped

Please note that this trait is only meant as speed-up helper. In some odd circumcises askama might still decide to HTML escape the input, so if this must not happen, then you need to use the |safe filter to prevent the auto escaping.

If you are unsure if your type generates HTML safe output in all cases, then DON’T mark it. Better safe than sorry!